Fixed bubble-sort and added insertion-sort.
Had a bug that was making the bubble-sort not be fully sorted and cause
some values to be incorrect. This was due to the setTimout being called after the
values in the array were swapped so I was not actually sorting the heights correctly.
Made the needed change and is now issue free.