Added merge-sort, insertion-sort, and fixed the animations.
Added the merge-sort and insertion-sort algorithms. The animations now is stored in
an array and is either a comparison or a chnage in the values of height. The animation
then is executed by iterating through the animations array and performing the correct change
to the bars based on the animation element. The timing is all synced using a variable called kft
that is incremented after every element inside of the animations array. This has resulted in
correct animations for the given algorithms.