Added quicksort and sound functionality.
Now the sorter will play sounds based on the current values being compared.
This uses the AudioContext API to play sounds at a certain frequency. The frequency
is determinded by the height of the current bar being compared. Swapping algorithms
like insertion, bubble, and quick only play one sound unlike merge-sort. This is because
these algorithms keep on value as a comparing value for a few iterations. This causes the
app to play certain pitches for extended times that sometimes sound like white noise. Only
one sound is played for these to account for this issue.