Having issues with express.js.
Currently trying to have request go to localhost:5000 from localhost:3000.
It was working previously but now it is not. Trying to log firstnames into
groups based on same lastnames. The other request will submit a last lastname
and will be given a list of firstnames that all share the same lastname. Some weird
stuff is happening so I'll try again later. Also, all the firstnames will be displayed
and updated using setState() from React but it is still weird.