Finished the simple-data-log.
This application takes a first and last name as input and can dislay
all first names that share a common last name. This is done by using
express.js to store the data within an object/map that stores the first
names into lists that correspond the last name keys. There is no data base
but I will soon create another project that will take in more complex data and
will use MongoDB as a backend.